Rename features2d test folder/namespace to features; move BOW tests to xfeatures2d
Test-side follow-up to the features2d->features rename and the BOW relocation:
- test/features2d -> test/features, namespace OpenCvSharp.Tests.Features2D ->
  OpenCvSharp.Tests.Features (Fast/Flann/ORB/SIFT tests).
- BOWImgDescriptorExtractorTest / BOWKMeansTrainerTest moved to test/xfeatures2d
  (namespace OpenCvSharp.Tests.XFeatures2D), matching the BOW move to
  cv::xfeatures2d.

Verified: test build + Features/XFeatures2D tests pass (36).
