+++ test/OpenCvSharp.Tests/features/FlannBasedMatcherTest.cs
@@ -1,7 +1,7 @@
 ﻿using OpenCvSharp.Flann;
 using Xunit;
 
-namespace OpenCvSharp.Tests.Features2D;
+namespace OpenCvSharp.Tests.Features;
 
 // ReSharper disable once InconsistentNaming
 public class FlannBasedMatcherTest : TestBase
